requirements.txt
astroid==1.6.0
colorama==0.3.9
diff-match-patch==20121119
Django==1.11.8
django-contrib-comments==1.8.0
django-crispy-forms==1.7.0
django-formtools==2.1
django-haystack==2.6.1
django-import-export==0.6.1
django-reversion==2.0.12
django-threadedcomments==1.1
et-xmlfile==1.0.1
flake8==3.5.0
future==0.16.0
httplib2==0.9.2
isort==4.2.15
jdcal==1.3
jieba==0.39
lazy-object-proxy==1.3.1
Markdown==2.6.10
mccabe==0.6.1
odfpy==1.3.6
openpyxl==2.4.9
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.1
PyMySQL==0.7.11
pytz==2017.3
PyYAML==3.12
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
Whoosh==2.7.4
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
